requirements.txt
# Core dependencies
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
tokenizers==0.15.0
spacy==3.7.2
textblob==0.17.1
pyannote.audio==3.1.1
webrtcvad==2.0.10
librosa==0.10.1
pydub==0.25.1
scipy==1.11.3
numpy==1.24.3
scikit-learn==1.3.0

# Backend framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# Task queue and messaging
celery==5.3.4
redis==5.0.1
paho-mqtt==1.6.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1

# Development and testing
pytest==7.4.3
httpx==0.25.0

# Download spaCy model
# Run: python -m spacy download en_core_web_sm